server/requirements.txt
fastapi==0.115.0
uvicorn==0.34.0
asyncpg==0.30.0
python-dotenv==1.0.1
pydantic==2.10.0
httpx==0.28.0
pytest==8.3.0
pytest-asyncio==0.25.0
